extract getValidatedMove function
