+++ __test__/cli/play.test.js
@@ -1,14 +1,14 @@
 const play = require('../../src/cli/play')
 const gameOver = require('../../src/cli/gameOver')
-const getMove = require('../../src/cli/getMove')
+const getMove = require('../../src/cli/getInput')
 const game = require('../../src/game')
 const referee = require('../../src/game/referee')
 
 jest.mock('../../src/game')
 jest.mock('../../src/cli/gameOver')
-jest.mock('../../src/cli/getMove')
+jest.mock('../../src/cli/getInput')
 
-describe('the makeMove function', () => {
+describe('the play function', () => {
   let options, write, io
   beforeEach(() => {
     jest.resetAllMocks()
