+++ src/cli/getInput.js
