+++ src/cli/play.js
@@ -1,12 +1,12 @@
 const drawBoard = require('./drawBoard')
-const getMove = require('./getMove')
+const getValidatedMove = require('./getValidatedMove')
 const gameOver = require('./gameOver')
 
-const { update, validator } = require('../game/')
+const { update } = require('../game/')
 
 const play = async (game, io) => {
   const { board, messages } = game
-  const { input, output, write } = io
+  const { write } = io
 
   if (gameIsOver(game)) {
     return gameOver(game, io)
@@ -14,13 +14,7 @@ const play = async (game, io) => {
 
   write(drawBoard(board))
 
-  let position = await getMove(input, output, messages.turn)
-  let validated = validator(position, board)
-
-  while (validated.status !== 'ok') {
-    position = await getMove(input, output, validated.message)
-    validated = validator(position, board)
-  }
+  const position = await getValidatedMove(board, messages.turn, io)
 
   play(update(position, game), io)
 }
